cloud_computing/api-skin-disease/requirements.txt
Flask
tensorflow
numpy
Pillow
gunicorn
google-cloud-storage